Cleanup, removed encoding from context, put context in it's own file.
