+++ src/Alchemy/Classes/UserContext.cs
@@ -1,31 +1,5 @@
-﻿/*
-Copyright 2011 Olivine Labs, LLC.
-http://www.olivinelabs.com
-*/
-
-/*
-This file is part of Alchemy Websockets.
-
-Alchemy Websockets is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-Alchemy Websockets is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License
-along with Alchemy Websockets.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-using System;
+﻿using System;
 using System.Net;
-using System.Net.Sockets;
-using System.Text;
-using System.Threading;
-using Alchemy.Handlers;
 using Alchemy.Handlers.WebSocket;
 
 namespace Alchemy.Classes
@@ -35,8 +9,6 @@ namespace Alchemy.Classes
     /// </summary>
     public class UserContext
     {
-        public readonly Header Header;
-
         /// <summary>
         /// AQ Link to the parent User Context
         /// </summary>
@@ -58,20 +30,21 @@ namespace Alchemy.Classes
         public DataFrame DataFrame;
 
         /// <summary>
-        /// What character encoding to use.
-        /// </summary>
-        public UTF8Encoding Encoding = new UTF8Encoding();
-
-        /// <summary>
         /// OnEvent Delegates specific to this connection.
         /// </summary>
         protected OnEventDelegate OnConnectDelegate = x => { };
-
         protected OnEventDelegate OnConnectedDelegate = x => { };
         protected OnEventDelegate OnDisconnectDelegate = x => { };
         protected OnEventDelegate OnReceiveDelegate = x => { };
         protected OnEventDelegate OnSendDelegate = x => { };
 
+        /// <summary>
+        /// The internal context connection header
+        /// </summary>
+        public Header Header
+        {
+            get { return _context.Header; }
+        }
         /// <summary>
         /// The type of connection this is
         /// </summary>
@@ -89,7 +62,6 @@ namespace Alchemy.Classes
         public UserContext(Context context)
         {
             _context = context;
-            Header = context.Header;
         }
 
         /// <summary>
@@ -248,140 +220,6 @@ namespace Alchemy.Classes
             DataFrame dataFrame = DataFrame.CreateInstance();
             dataFrame.Append(someBytes);
             _context.Handler.Send(dataFrame, _context, raw, close);
-        }
-    }
-
-    /// <summary>
-    /// This class contains the required data for each connection to the server.
-    /// </summary>
-    public class Context : IDisposable
-    {
-        /// <summary>
-        /// The exported version of this context.
-        /// </summary>
-        public readonly UserContext UserContext;
-
-        /// <summary>
-        /// The buffer used for accepting raw data from the socket.
-        /// </summary>
-        public byte[] Buffer;
-
-        /// <summary>
-        /// Whether or not the TCPClient is still connected.
-        /// </summary>
-        public bool Connected = true;
-
-        /// <summary>
-        /// The raw client connection.
-        /// </summary>
-        public TcpClient Connection;
-
-        /// <summary>
-        /// The current connection handler.
-        /// </summary>
-        public Handler Handler = Handler.Instance;
-
-        /// <summary>
-        /// The Header
-        /// </summary>
-        public Header Header;
-
-        /// <summary>
-        /// Whether or not this client has passed all the setup routines for the current handler(authentication, etc)
-        /// </summary>
-        public Boolean IsSetup;
-
-        /// <summary>
-        /// How many pings in a row we've had from this client, indicates inactivity.
-        /// </summary>
-        public int Pings;
-
-        /// <summary>
-        /// Semaphores that limit sends and receives to 1 and a time.
-        /// </summary>
-        public SemaphoreSlim ReceiveReady = new SemaphoreSlim(1);
-
-        /// <summary>
-        /// How many bytes we received this tick.
-        /// </summary>
-        public int ReceivedByteCount;
-
-        public SemaphoreSlim SendReady = new SemaphoreSlim(1);
-
-        /// <summary>
-        /// A link to the server listener instance this client is currently hosted on.
-        /// </summary>
-        public WebSocketServer Server;
-
-        private int _bufferSize = 512;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="Context"/> class.
-        /// </summary>
-        public Context()
-        {
-            Buffer = new byte[_bufferSize];
-            UserContext = new UserContext(this);
-        }
-
-        /// <summary>
-        /// Gets or sets the size of the buffer.
-        /// </summary>
-        /// <value>
-        /// The size of the buffer.
-        /// </value>
-        public int BufferSize
-        {
-            get { return _bufferSize; }
-            set
-            {
-                _bufferSize = value;
-                Buffer = new byte[_bufferSize];
-            }
-        }
-
-        #region IDisposable Members
-
-        /// <summary>
-        /// Performs application-defined tasks associated with freeing, releasing, or resetting unmanaged resources.
-        /// </summary>
-        public void Dispose()
-        {
-            try
-            {
-                Connection.Client.Close();
-                Connection = null;
-            }
-            catch (Exception e)
-            {
-                Server.Log.Debug("Client Already Disconnected", e);
-            }
-            finally
-            {
-                if (Connected)
-                {
-                    Connected = false;
-                }
-                UserContext.OnDisconnect();
-            }
-        }
-
-        #endregion
-
-        /// <summary>
-        /// Resets this instance.
-        /// Clears the dataframe if necessary. Resets Received byte count.
-        /// </summary>
-        public void Reset()
-        {
-            if (UserContext.DataFrame != null)
-            {
-                if (UserContext.DataFrame.State == DataFrame.DataState.Complete)
-                {
-                    UserContext.DataFrame.Reset();
-                }
-            }
-            ReceivedByteCount = 0;
         }
     }
 }
