+++ src/Alchemy/Handlers/Handler.cs
@@ -1,27 +1,6 @@
-﻿/*
-Copyright 2011 Olivine Labs, LLC.
-http://www.olivinelabs.com
-*/
-
-/*
-This file is part of Alchemy Websockets.
-
-Alchemy Websockets is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-Alchemy Websockets is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License
-along with Alchemy Websockets.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-using System;
+﻿using System;
 using System.Net.Sockets;
+using System.Text;
 using System.Threading;
 using Alchemy.Classes;
 using Alchemy.Handlers.WebSocket;
@@ -78,7 +57,7 @@ namespace Alchemy.Handlers
         /// <param name="context">The user context.</param>
         public void ProcessHeader(Context context)
         {
-            string data = context.UserContext.Encoding.GetString(context.Buffer, 0, context.ReceivedByteCount);
+            string data = Encoding.UTF8.GetString(context.Buffer, 0, context.ReceivedByteCount);
             //Check first to see if this is a flash socket XML request.
             if (data == "<policy-file-request/>\0")
             {
