+++ src/Alchemy/Handlers/WebSocket/hybi00/DataFrame.cs
@@ -1,26 +1,4 @@
-﻿/*
-Copyright 2011 Olivine Labs, LLC.
-http://www.olivinelabs.com
-*/
-
-/*
-This file is part of Alchemy Websockets.
-
-Alchemy Websockets is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-Alchemy Websockets is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License
-along with Alchemy Websockets.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-using System;
+﻿using System;
 using System.Collections.Generic;
 
 namespace Alchemy.Handlers.WebSocket.hybi00
@@ -52,8 +30,8 @@ namespace Alchemy.Handlers.WebSocket.hybi00
                 // wrap the array with the wrapper bytes
                 var startBytes = new byte[1];
                 var endBytes = new byte[1];
-                Payload.Insert(0, new ArraySegment<byte>(startBytes));
-                Payload.Add(new ArraySegment<byte>(endBytes));
+                Payload.Insert(0, new ArraySegment<byte>(startBytes));//Add header byte
+                Payload.Add(new ArraySegment<byte>(endBytes));//put termination byte at end
                 Format = DataFormat.Frame;
             }
             return Payload;
@@ -66,8 +44,8 @@ namespace Alchemy.Handlers.WebSocket.hybi00
         {
             if (Format == DataFormat.Frame)
             {
-                Payload.RemoveAt(0);
-                Payload.RemoveAt(Payload.Count - 1);
+                Payload.RemoveAt(0);//remove header byte
+                Payload.RemoveAt(Payload.Count - 1);//remove termination byte
                 Format = DataFormat.Raw;
             }
             return Payload;
