+++ src/Alchemy/Handlers/WebSocket/hybi10/Authentication.cs
@@ -1,27 +1,6 @@
-﻿/*
-Copyright 2011 Olivine Labs, LLC.
-http://www.olivinelabs.com
-*/
-
-/*
-This file is part of Alchemy Websockets.
-
-Alchemy Websockets is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-Alchemy Websockets is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License
-along with Alchemy Websockets.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-using System;
+﻿using System;
 using System.Security.Cryptography;
+using System.Text;
 using Alchemy.Classes;
 
 namespace Alchemy.Handlers.WebSocket.hybi10
@@ -92,7 +71,7 @@ namespace Alchemy.Handlers.WebSocket.hybi10
         {
             // generate a byte array representation of the handshake including the answer to the challenge
             string temp = handshake.ToString();
-            byte[] handshakeBytes = context.UserContext.Encoding.GetBytes(temp);
+            byte[] handshakeBytes = Encoding.UTF8.GetBytes(temp);
             context.UserContext.Send(handshakeBytes, true);
         }
 
@@ -102,7 +81,7 @@ namespace Alchemy.Handlers.WebSocket.hybi10
 
             // Create a hash of the rawAnswer and return it
             SHA1 hasher = SHA1.Create();
-            return Convert.ToBase64String(hasher.ComputeHash(context.UserContext.Encoding.GetBytes(rawAnswer)));
+            return Convert.ToBase64String(hasher.ComputeHash(Encoding.UTF8.GetBytes(rawAnswer)));
         }
     }
 }
