+++ src/Alchemy/Handlers/WebSocket/hybi10/DataFrame.cs
@@ -1,26 +1,4 @@
-﻿/*
-Copyright 2011 Olivine Labs, LLC.
-http://www.olivinelabs.com
-*/
-
-/*
-This file is part of Alchemy Websockets.
-
-Alchemy Websockets is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-Alchemy Websockets is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License
-along with Alchemy Websockets.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-using System;
+﻿using System;
 using System.Collections.Generic;
 
 namespace Alchemy.Handlers.WebSocket.hybi10
@@ -66,12 +44,12 @@ namespace Alchemy.Handlers.WebSocket.hybi10
                 switch (State)
                 {
                     case DataState.Pong:
-                        _header.OpCode = OpCode.Pong;
+                        _header.OpCode = OpCode.Pong;//Setup Opcode for Pong frame if application has specified that we're sending a pong.
                         break;
                 }
                 byte[] headerBytes = _header.ToBytes();
-                Mask();
-                Payload.Insert(0, new ArraySegment<byte>(headerBytes));
+                Mask();//Uses _header, must call ToBytes before calling Mask
+                Payload.Insert(0, new ArraySegment<byte>(headerBytes));//put header at first position
                 Format = DataFormat.Frame;
             }
             return Payload;
@@ -84,8 +62,8 @@ namespace Alchemy.Handlers.WebSocket.hybi10
         {
             if (Format == DataFormat.Frame)
             {
-                Payload.RemoveAt(0);
-                Mask();
+                Payload.RemoveAt(0);//Remove header bytes
+                Mask();//unmask data
                 Format = DataFormat.Raw;
             }
             return Payload;
